Add skill publish UI with OAuth login and gateway token storage auth manager
Introduces a self-service skill submission flow: a publish router serving
the form and proxying publishes to mcp-server-gateway, a SkillAuthService
that registers the Okta OAuth provider and initiates login, and a
GatewayTokenStorageAuthManager that auto-registers MCP OAuth providers
from plugin configs on callback (handling server-restart scenarios).
