+++ language_model_gateway/gateway/routers/skill_publish_router.py
@@ -1,28 +1,21 @@
-import logging
 import os
 from enum import Enum
 from pathlib import Path
 from typing import Annotated, Sequence
 
-import httpx
 from fastapi import APIRouter, Depends, HTTPException, Query, Request, params
-from fastapi.responses import FileResponse, JSONResponse, RedirectResponse
+from fastapi.responses import FileResponse
 from oidcauthlib.auth.fastapi_auth_manager import FastAPIAuthManager
 from simple_container.container.inject import Inject
 from starlette.responses import Response
 
 from languagemodelcommon.auth.oauth_provider_registrar import OAuthProviderRegistrar
-from languagemodelcommon.configs.schemas.config_schema import McpOAuthConfig
-from language_model_gateway.gateway.utilities.logger.log_levels import SRC_LOG_LEVELS
-
-logger = logging.getLogger(__name__)
-logger.setLevel(SRC_LOG_LEVELS["AUTH"])
+from language_model_gateway.gateway.skills.skill_auth_service import SkillAuthService
+from language_model_gateway.gateway.skills.skill_publish_client import (
+    SkillPublishClient,
+)
 
 _STATIC_DIR = Path(__file__).resolve().parents[2] / "static"
-
-_SKILLS_PUBLISHER_CLIENT_ID = "0oa11g45c90Fqbgzz698"
-_SKILLS_PUBLISHER_AUTH_PROVIDER = f"mcp_oauth_{_SKILLS_PUBLISHER_CLIENT_ID}"
-_OKTA_METADATA_URL = "https://icanbwell.okta.com/.well-known/openid-configuration"
 
 
 class SkillPublishRouter:
@@ -43,8 +36,14 @@ class SkillPublishRouter:
             tags=self.tags,
             dependencies=self.dependencies,
         )
-        self._mcp_server_gateway_url = os.environ.get(
+        mcp_server_gateway_url = os.environ.get(
             "MCP_SERVER_GATEWAY_URL", "http://mcp_server_gateway:5000"
+        )
+        self._auth_service = SkillAuthService(
+            mcp_server_gateway_url=mcp_server_gateway_url
+        )
+        self._publish_client = SkillPublishClient(
+            mcp_server_gateway_url=mcp_server_gateway_url
         )
         self._register_routes()
 
@@ -86,27 +85,13 @@ class SkillPublishRouter:
         ],
         return_url: str = Query(default="/skills/publish"),
     ) -> Response:
-        """Initiate OAuth login for the skills-publisher.
-
-        Redirects to Okta with state containing return_url so that the
-        callback handler renders the sessionStorage template.
-        """
-        await self._ensure_provider_registered(
+        """Initiate OAuth login for the skills-publisher."""
+        return await self._auth_service.initiate_login(
+            request=request,
             auth_manager=auth_manager,
             oauth_provider_registrar=oauth_provider_registrar,
+            return_url=return_url,
         )
-
-        redirect_uri = self._get_auth_callback_uri(request)
-
-        url = await auth_manager.create_authorization_url(
-            auth_provider=_SKILLS_PUBLISHER_AUTH_PROVIDER,
-            redirect_uri=redirect_uri,
-            url=return_url,
-            referring_email="skill-submit-ui",
-            referring_subject="skill-submit-ui",
-        )
-
-        return RedirectResponse(url, status_code=302)
 
     async def publish_skill(self, request: Request) -> Response:
         """Proxy the publish request to the mcp-server-gateway REST API."""
@@ -118,57 +103,7 @@ class SkillPublishRouter:
             )
 
         body = await request.json()
-        rest_url = f"{self._mcp_server_gateway_url}/api/skills/publish"
-        headers = {
-            "Authorization": auth_header,
-            "Content-Type": "application/json",
-        }
-
-        async with httpx.AsyncClient(timeout=60.0) as client:
-            try:
-                response = await client.post(rest_url, json=body, headers=headers)
-            except httpx.HTTPError as exc:
-                return JSONResponse(
-                    status_code=502,
-                    content={"error": f"Network error: {exc}"},
-                )
-
-        try:
-            content = response.json()
-        except Exception:
-            content = {"error": response.text or f"HTTP {response.status_code}"}
-
-        return JSONResponse(
-            status_code=response.status_code,
-            content=content,
-        )
-
-    async def _ensure_provider_registered(
-        self,
-        *,
-        auth_manager: FastAPIAuthManager,
-        oauth_provider_registrar: OAuthProviderRegistrar,
-    ) -> None:
-        """Ensure the skills-publisher OAuth provider is registered."""
-        oauth_config = McpOAuthConfig(
-            authServerMetadataUrl=_OKTA_METADATA_URL,
-            clientId=_SKILLS_PUBLISHER_CLIENT_ID,
-            displayName="Okta b.well",
-        )
-        await oauth_provider_registrar.register_provider(
-            auth_provider=_SKILLS_PUBLISHER_AUTH_PROVIDER,
-            oauth=oauth_config,
-            server_url=f"{self._mcp_server_gateway_url}/skills-publisher/",
-            auth_manager=auth_manager,
-        )
-
-    @staticmethod
-    def _get_auth_callback_uri(request: Request) -> str:
-        """Return the redirect_uri for the OAuth callback (existing endpoint)."""
-        auth_redirect_uri = os.environ.get("AUTH_REDIRECT_URI")
-        if auth_redirect_uri:
-            return auth_redirect_uri
-        return str(request.url_for("auth_callback"))
+        return await self._publish_client.publish(body=body, auth_header=auth_header)
 
     def get_router(self) -> APIRouter:
         return self.router
